Convert to noarch Python (#9)
* Make the package `noarch: python`

As this is a pure Python package with no selectors or use of 2to3, go
ahead and make it `noarch: python`. Should cutdown the maintenance
burden here drastically.

* MNT: Re-rendered with conda-smithy 3.1.12 and pinning 2018.11.27

As the package is now a `noarch: python` package, re-render to update
the CI configurations and other repo content accordingly.

* Bump build number to 1002

As the package is now `noarch: python`, bump the build number to `1002`
to create new packages.